Thicker border and border on *both* pages.
Explicitly printing a border around the who page and then cutting it out
does solve the problem of margins in a printer-agnostic way, but
unfortunately it's also a bit cumbersome and it leads to slight
downscaling of the page.

Also, I tink that having three seals might be somewhat of an overkill.
Perhaps I should remove this in the future.